Create AbstractMutableBeaconStateImpl to hold common logic
